+++ lib/SSIEth/src/PHY.cpp
@@ -1,5 +1,5 @@
 #include "PHY.h"
-#include <utils_assert.h>
+#include <utils/utils_assert.h>
 #include "ieee8023_mii_standard_register.h"
 
 /**
